requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
markdownify==0.10.2
nltk==3.8.1
PyYAML==6.0
urllib3==1.26.15
